メソッドを移動
